Dockerfile: Convert to Ubuntu multi-stage
The main thing needed is to move to Ubuntu for the container. We
are about to add support for authenticating to the device gateway
with an HSM and Alpine has some problems:

 * missing libp11
 * crashes when its built and we try to use it

This gets us on Ubuntu which is probably a little more friendly to build
a general purpose testing container with. It also makes it a multi-stage
build so that things are faster to build. We are also able to build all
of LTP under Ubuntu

Signed-off-by: Andy Doan <andy@foundries.io>