Move to 19.10
Ubuntu 20.04 won't build on armhf. This is because our CI uses
docker:dind which is currently at 19.03.8 for armhf. The fix is at
19.03.9 but they stopped building docker:dind for armhf.

Eventually - we have to fix this in CI because nobody is going to be
capable of doing ubuntu:20.04 containers for armhf. For now this gets us
going.

Why not use the 18.04 LTS? because its pycurl/libcurl wasn't compatible
with how this code was written.

Signed-off-by: Andy Doan <andy@foundries.io>